Unwrap data on successful calls
This makes the library a little more ergonomic to use. Since the Ordermentum API uses an envelope response, there's no need to have access to headers in a successful case.

The only time you may need access to the specific status code, etc, is on failure. In that case, it's all available from the thrown error.